Rename job util module / job test module, add `test_utils` module for testing job utils and remove useless code in job tests (#17996)
