+++ mlflow/server/job_api.py
@@ -74,7 +74,7 @@ class SubmitJobPayload(BaseModel):
 @job_api_router.post("/", response_model=Job)
 def submit_job(payload: SubmitJobPayload) -> Job:
     from mlflow.server.jobs import submit_job
-    from mlflow.server.jobs.util import _load_function
+    from mlflow.server.jobs.utils import _load_function
 
     function_fullname = payload.function_fullname
     try:
