+++ mlflow/server/jobs/__init__.py
@@ -95,7 +95,7 @@ def submit_job(
         the updated job entity.
     """
     from mlflow.environment_variables import MLFLOW_SERVER_ENABLE_JOB_EXECUTION
-    from mlflow.server.jobs.util import (
+    from mlflow.server.jobs.utils import (
         _check_requirements,
         _get_or_init_huey_instance,
         _validate_function_parameters,
