+++ mlflow/server/jobs/_huey_consumer.py
@@ -18,7 +18,7 @@ import os
 import threading
 
 from mlflow.server import MLFLOW_HUEY_INSTANCE_KEY
-from mlflow.server.jobs.util import (
+from mlflow.server.jobs.utils import (
     _exit_when_orphaned,
     _get_or_init_huey_instance,
 )
