+++ tests/server/jobs/test_utils.py
@@ -1,0 +1,59 @@
+import os
+
+import pytest
+
+from mlflow.exceptions import MlflowException
+from mlflow.server.jobs.utils import _validate_function_parameters
+
+pytestmark = pytest.mark.skipif(
+    os.name == "nt", reason="MLflow job execution is not supported on Windows"
+)
+
+
+def test_validate_function_parameters():
+    """Test the parameter validation function directly."""
+
+    def test_func(a, b, c=None):
+        return a + b + (c or 0)
+
+    # Test with all required parameters present
+    _validate_function_parameters(test_func, {"a": 1, "b": 2})
+    _validate_function_parameters(test_func, {"a": 1, "b": 2, "c": 3})
+
+    # Test with missing required parameters
+    with pytest.raises(MlflowException, match=r"Missing required parameters.*\['b'\]"):
+        _validate_function_parameters(test_func, {"a": 1})
+
+    assert True  # If we get here, the exception was properly raised
+
+    # Test with multiple missing required parameters
+    with pytest.raises(MlflowException, match=r"Missing required parameters.*\['a', 'b'\]"):
+        _validate_function_parameters(test_func, {})
+
+
+def test_validate_function_parameters_with_varargs():
+    """Test parameter validation with functions that have **kwargs."""
+
+    def test_func_with_kwargs(a, **kwargs):
+        return a
+
+    # Should not raise error even with extra parameters due to **kwargs
+    _validate_function_parameters(test_func_with_kwargs, {"a": 1, "extra": 2})
+
+    # Should still raise error for missing required parameters
+    with pytest.raises(MlflowException, match=r"Missing required parameters.*\['a'\]"):
+        _validate_function_parameters(test_func_with_kwargs, {"extra": 2})
+
+
+def test_validate_function_parameters_with_positional_args():
+    """Test parameter validation with functions that have *args."""
+
+    def test_func_with_args(a, *args):
+        return a
+
+    # Should work fine with just required parameter
+    _validate_function_parameters(test_func_with_args, {"a": 1})
+
+    # Should still raise error for missing required parameters
+    with pytest.raises(MlflowException, match=r"Missing required parameters.*\['a'\]"):
+        _validate_function_parameters(test_func_with_args, {})
